requirements.txt
Flask==3.0.2
tiktoken==0.6.0
langchain-community==0.0.25
langchain-openai==0.0.8
langchain==0.1.10
python-dotenv==1.0.1
faiss-cpu
openai
Flask-SQLAlchemy==3.1.1
